Move commonControllers.js and commonServices.js into a new folder called 'common'. Also rename the new source files added by me when creating the build system to be correct given what is inside them.
