+++ WebContent/js/appDirectivesModule.js
@@ -1,0 +1,2 @@
+import angular from 'angular';
+export default angular.module('appDirectives', []);
